修复强类型问题

+++ src/Components/swoole/src/Pool/BaseAsyncPool.php
@@ -325,7 +325,7 @@ abstract class BaseAsyncPool extends BasePool
     {
         if ((null !== Worker::getWorkerId() || Coroutine::stats()['coroutine_num'] > 0) && null !== ($heartbeatInterval = $this->config->getHeartbeatInterval()))
         {
-            $this->heartbeatTimerId = \Swoole\Timer::tick($heartbeatInterval * 1000, [$this, 'heartbeat']);
+            $this->heartbeatTimerId = \Swoole\Timer::tick((int) ($heartbeatInterval * 1000), [$this, 'heartbeat']);
             Event::on(['IMI.MAIN_SERVER.WORKER.EXIT', 'IMI.PROCESS.END'], function () {
                 $this->stopHeartbeat();
             }, \Imi\Util\ImiPriority::IMI_MIN);
